rtu layer: fix parsing of settings string
Throw exceptions on error instead of silently changing values.
Check settings string for given stop bit value.

Up until now, the stop bits value was hard coded to be 2 if the parity
mode was set to none or 1 if the parity mode was set to even or odd.

So a setting of "115200E1" would result in a stop bit value of 1.
But "115200E2" would result also in a stop bit value of 1.
And "115200N1" would result in a stop bit value of 2.

This is not a restriction of libmodbus. The library supports all these
modes just fine. In fact they are using 115200N1 in the documentation
for modbus_new_rtu().

Rather this is a bug in the string parsing that libmodbuspp introduces
in its constructor.
libmodbus has no string parsing in modbus_new_rtu() and takes in the
arguments as parameters directly.

Fixes: b0f3dc4be5e82729825158ca651657afe1f6380b